requirements/default.txt
data
ninja
accelerate>=0.20.3
einops==0.6.1
ftfy==6.1.1
iopath==0.1.10
ipdb==0.13.13
numpy==1.24.3
Pillow==9.5.0
PyYAML==6.0.1
regex==2022.10.31
pytorchvideo
fvcore
decord==0.6.0
tqdm
setuptools==65.5.1
bigmodelvis
nltk
tensorboard
cython
plyfile
trimesh
sentencepiece
